+++ packages/@leight/sdk/src/generator/generatorSourceApi.ts
@@ -1,3 +1,4 @@
+import {IPackageType}    from "@leight/generator";
 import {withSourceFile}  from "@leight/generator-server";
 import {normalize}       from "node:path";
 import {type IGenerator} from "../api";
@@ -12,11 +13,32 @@ export interface IGeneratorSourceApiParams {
      *
      * Defaults to './entity-schema'
      */
-    sdk?: string;
+    entitySchemaPackage?: string;
+    /**
+     * Another extension of generated ISource
+     *
+     * IFooSource extends ISource, sourceEx[]
+     */
+    sourceEx?: {
+        extends?: IPackageType[];
+    };
 }
 
-export const generatorSourceApi: IGenerator<IGeneratorSourceApiParams> = async ({packageName, file, barrel, params: {modelName, sdk = "./entity-schema"}}) => {
+export const generatorSourceApi: IGenerator<IGeneratorSourceApiParams> = async (
+    {
+        packageName,
+        file,
+        barrel,
+        params: {
+                    modelName,
+                    entitySchemaPackage = "./entity-schema",
+                    sourceEx,
+                }
+    }) => {
     withSourceFile()
+        .withHeader(`
+    Source code of the common stuff for ${modelName} which could be shared between server and client side.
+        `)
         .withImports({
             imports: {
                 "@leight/source":      [
@@ -27,7 +49,7 @@ export const generatorSourceApi: IGenerator<IGeneratorSourceApiParams> = async (
                 "@leight/react-query": [
                     "type IUseQuery",
                 ],
-                [sdk]:                 [
+                [entitySchemaPackage]: [
                     `type I${modelName}CreateSchema`,
                     `type I${modelName}FilterSchema`,
                     `type I${modelName}ParamSchema`,
@@ -37,6 +59,17 @@ export const generatorSourceApi: IGenerator<IGeneratorSourceApiParams> = async (
                 ]
             }
         })
+        .withImports(sourceEx?.extends ? {
+            imports: sourceEx.extends
+                         .filter(((item): item is Required<IPackageType> => Boolean(item.package)))
+                         .reduce((prev, {type, package: $package}) => ({
+                             ...prev,
+                             [$package]: [
+                                 `type ${type}`,
+                                 ...(prev[$package] || [])
+                             ],
+                         }), {} as Record<string, any>),
+        } : undefined)
         .withTypes({
             exports: {
                 [`IUse${modelName}FetchQuery`]: `IUseQuery<I${modelName}SourceSchema["Query"], I${modelName}SourceSchema["Entity"]>`,
@@ -46,10 +79,14 @@ export const generatorSourceApi: IGenerator<IGeneratorSourceApiParams> = async (
         .withInterfaces({
             exports: {
                 [`I${modelName}Source`]:       {
-                    extends: `ISource<I${modelName}SourceSchema>`,
+                    extends: [
+                                 {type: `ISource<I${modelName}SourceSchema>`},
+                             ].concat(sourceEx?.extends || []),
                 },
                 [`I${modelName}SourceSchema`]: {
-                    extends: `
+                    extends: [
+                        {
+                            type: `
 ISourceSchema<
     I${modelName}Schema,
     I${modelName}CreateSchema,
@@ -58,7 +95,9 @@ ISourceSchema<
     I${modelName}SortSchema,
     I${modelName}ParamSchema
  >
-                    `
+                            `,
+                        },
+                    ],
                 }
             }
         })