+++ packages/@leight/user/src/api/index.ts
@@ -4,3 +4,4 @@ export * from "./ITokenService";
 export * from "./IUserJwtService";
 export * from "./IUserService";
 export * from "./IUserSession";
+export * from "./IUserSourceEx";