+++ packages/@leight/user/src/leight-sdk.ts
@@ -6,4 +6,9 @@ import {
 void withSdk(withSourceGenerators({
     modelName:    "User",
     PrismaSchema: "@leight/prisma",
+    sourceEx:     {
+        extends: [
+            {type: "IUserSourceEx", package: "../api"},
+        ],
+    },
 }));