+++ packages/@leight/generator-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/generator-server
 
+## 0.5.2
+
+### Patch Changes
+
+- Improved generator
+- Improved generator, enabled extensions, fixed UserSource
+- Updated dependencies
+    - @leight/generator@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes