+++ packages/@leight/generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/generator
 
+## 0.5.2
+
+### Patch Changes
+
+- Improved generator, enabled extensions, fixed UserSource
+
 ## 0.5.1
 
 ### Patch Changes