+++ packages/@leight/job-server/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leight/job-server
 
+## 0.5.50
+
+### Patch Changes
+
+- Improved generator
+- Improved generator, enabled extensions, fixed UserSource
+- Updated dependencies
+- Updated dependencies
+    - @leight/user@0.5.51
+    - @leight/job@0.5.51
+
 ## 0.5.49
 
 ### Patch Changes