+++ packages/@leight/job/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @leight/job
 
+## 0.5.51
+
+### Patch Changes
+
+- Improved generator
+- Improved generator, enabled extensions, fixed UserSource
+
 ## 0.5.50
 
 ### Patch Changes