+++ packages/@leight/sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leight/sdk
 
+## 0.5.2
+
+### Patch Changes
+
+- Improved generator
+- Improved generator, enabled extensions, fixed UserSource
+- Updated dependencies
+- Updated dependencies
+    - @leight/generator-server@0.5.2
+    - @leight/generator@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes